requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr<2.0,>=1.4
argparse
cmd2>=0.6.7
PrettyTable<0.8,>=0.7
pyparsing>=2.0.1
six>=1.9.0
stevedore>=1.5.0 # Apache-2.0
unicodecsv>=0.8.0
PyYAML>=3.1.0
